backend/requirements.txt
# Core FastAPI and server - use compatible versions
fastapi==0.115.4
uvicorn[standard]==0.32.1
python-dotenv==1.0.1

# Database
pymongo==4.6.0

# Google AI APIs - pin to working versions
google-genai==1.35.0
Pillow==10.4.0
pydantic==2.10.3
python-multipart==0.0.12
aiofiles==24.1.0

# Google Cloud Storage for video persistence
google-cloud-storage==2.10.0

# Force compatible anyio version
anyio==4.8.0
